agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-agent"
version = "0.1.0"
description = "A Strands AI agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.14.0",
    "mypy>=1.18.0",
    "black>=25.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
